app/requirements.txt
# WPH Pharmacy SMART - Python Dependencies

# Web Framework
flask==3.0.3
waitress==3.0.0

# Database
psycopg2-binary==2.9.9

# Data Processing
pandas==2.2.2
openpyxl==3.1.5

# API & HTTP
requests==2.31.0

# Configuration
python-dotenv==1.0.0

# System Monitoring (NEW)
psutil==5.9.8

# Validation
pydantic==2.12.1
pydantic-settings==2.5.2

# Optional: for future ML/AI features
# scikit-learn==1.3.0
# numpy==1.24.3
